GTK: Add reload button to list-view-page, fix layout
